backout this for now

+++ servers/slapd/back-ldap/bind.c
@@ -175,7 +175,7 @@ ldap_back_conn_cmp(
 	/* For shared sessions, conn is NULL. Only explicitly
 	 * bound sessions will have non-NULL conn.
 	 */
-	return (int)lc1->conn - (int)lc2->conn;
+	return lc1->conn - lc2->conn;
 }
 
 /*
